### 2023-11-20 #6070 @editorialbot
[PRE REVIEW]: dropout: an R Package for Addressing Dropouts, Missing Values, and Sectional Challenges in Survey Data Analysis
**Submitting author:** @hendr1km (<a href="http://orcid.org/0009-0007-3222-899X">Hendrik Mann</a>)
**Repository:** https://github.com/hendr1km/dropout
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.0
