+++ articles/paper.md
@@ -1,92 +1,48 @@
+---
+title: 'dropout: an R Package for Addressing Dropouts, Missing Values, and Sectional
+  Challenges in Survey Data Analysis'
+tags:
+- R
+- dropout
+- missing values
+- participation
+- survey data
+date: "19 November 2023"
+output:
+  pdf_document: default
+  word_document: default
+bibliography: paper.bib
+editor_options:
+  chunk_output_type: console
+authors:
+- name: Hendrik Mann
+  orcid: "0009-0007-3222-899X"
+  affiliation: University of Wuppertal, Germany
+  output: md_document
+---
+
 # Summary
 
-Missing data can reduce the statistical power of a study and can produce
-biased estimates, leading to invalid conclusions (Kang 2013). When
-working with survey data, capturing and categorizing missing values is
-crucial for maintaining data quality. The `dropout` package assists in
-distinguishing whether missing values are isolated instances, sectional
-missing values, or complete dropouts. This distinction enables effective
-data cleaning and provides insights into the study design and the
-response behaviors of participants.
+Missing data can reduce the statistical power of a study and can produce biased estimates, leading to invalid conclusions [@Kang2013]. When working with survey data, capturing and categorizing missing values is crucial for maintaining data quality. The `dropout` package assists in distinguishing whether missing values are isolated instances, sectional missing values, or complete dropouts. This distinction enables effective data cleaning and provides insights into the study design and the response behaviors of participants.
 
 # Statement of need
 
-`dropout` is an R package (R Core Team 2022) available on CRAN, designed
-to differentiate various types of missing values in survey data. It
-enables users to identify whether missing values are due to complete
-dropouts—participants ceasing to answer the questionnaire entirely,
-participants skipping entire sections, or isolated instances of ‘NA’
-values. This distinction is crucial for accurate data analysis and
-understanding participant engagement in surveys. Beyond creating summary
-statistics, the package adeptly identifies participants who have either
-completely dropped out, skipped sections, or left specific questions
-unanswered. This capability not only facilitates thorough data cleaning
-but also provides insights into the various factors influencing dropout
-behavior. The `dropout` package’s scalability is enhanced by the
-integration of C++ code through the Rcpp package, developed by
-Eddelbuettel et al. (2023). This incorporation makes `dropout`
-particularly effective for analyzing large survey datasets.
+`dropout` is an R package [@R] available on CRAN, designed to differentiate various types of missing values in survey data. It enables users to identify whether missing values are due to complete dropouts—participants ceasing to answer the questionnaire entirely, participants skipping entire sections, or isolated instances of 'NA' values. This distinction is crucial for accurate data analysis and understanding participant engagement in surveys.
+Beyond creating summary statistics, the package adeptly identifies participants who have either completely dropped out, skipped sections, or left specific questions unanswered. This capability not only facilitates thorough data cleaning but also provides insights into the various factors influencing dropout behavior.
+The `dropout` package's scalability is enhanced by the integration of C++ code through the Rcpp package, developed by @rcpp. This incorporation makes `dropout` particularly effective for analyzing large survey datasets.
 
 # Usage
 
-`dropout` includes two essential functions, beginning with
-`drop_summary`. This function is designed to generate summary statistics
-for missing values in your dataset. It requires two primary inputs: the
-dataframe to analyze and the identifier of the last survey item. The
-latter is particularly important in cases where your data includes
-additional columns generated by the survey platform, such as
-participation time. These extra items can lead to biased results in the
-function, as no dropouts may be detected if the last column has no
-missing values. To counter this, if the last\_col argument is left
-unspecified, `drop_summary` will automatically address the issue by
-setting `last_col` to the last column in the dataframe that contains
-missing values, and it will issue a warning to inform the user. For
[…]
