### 2023-11-20 #6070 @hendr1km
Am 20.11.2023 um 18:38 schrieb The Open Journals editorial robot ***@***.***>:

⚠️ An error happened when generating the pdf<https://github.com/openjournals/joss-papers/actions/runs/6933760887?check_suite_focus=true>.
