### 2023-11-20 #6070 @hendr1km
Am 20.11.2023 um 21:01 schrieb The Open Journals editorial robot ***@***.***>:

Hello @hendr1km<https://github.com/hendr1km>, here are the things you can ask me to do:
