+++ articles/paper.md
@@ -1,17 +1,18 @@
 ---
 title: 'dropout: an R Package for Addressing Dropouts, Missing Values, and Sectional Challenges in Survey Data Analysis'
 tags:
-- R
+- r
 - dropout
 - missing values
 - participation
 - survey data
+date: "19 November 2023"
+output: pdf_document
+bibliography: paper.bib
 authors:
 - name: Hendrik Mann
-  orcid: 0009-0007-3222-899X
-  affiliation: University of Wuppertal
-date: "19 November 2023"
-bibliography: paper.bib
+  orcid: "0009-0007-3222-899X"
+  affiliation: "University of Wuppertal, Germany"
 ---
 
 # Summary
@@ -36,88 +37,102 @@ For the following examples we will use an adapted version of the Flying Etiquett
 
 As illustrated with the `flying` dataset example, even though all columns are arranged in the correct order of survey items, the last column survey_type does not correspond to a survey item. In such scenarios, the dropout package intuitively addresses this issue by disregarding the non-survey column and automatically setting the last_col argument to `location_census_region`. This adjustment is accompanied by a warning to inform the user. However, in more complex situations, it's advisable to either create a subset of your data or manually set the last_col argument to the actual last survey item. We will demonstrate this approach in the following examples.
 
+```r
+# install.packages(c("dropout", "tidyverse"))
 
-    # install.packages(c("dropout", "tidyverse"))
-
-    library(dropout)
-    library(tidyverse)
-    data(flying)
-
+library(dropout)
+library(tidyverse)
+data(flying)
+```
 
 Initially, we use the `drop_summary` function to generate an overview of the different types of missing values in our dataframe. From this analysis, it becomes evident that certain parts of the survey experience higher dropout rates. Notably, 18 participants dropped out early, at the third survey item, culminating in a total of 42 dropouts by the end of the survey. The ´section_na´ column reveals that 164 participants skipped an entire section of the questionnaire, or at least a consecutive portion identifiable as a section in this context. Furthermore, single missing values are particularly prevalent in responses to the household income question. 
 
 
-    flying %>% 
-    drop_summary(last_col = "location_census_region") %>% 
-    print(n = Inf) 
+```r
+flying %>% 
+drop_summary(last_col = "location_census_region") %>% 
+print(n = Inf) 
+```
+
 
 In the subsequent step, we aim to refine our dataset to include only those survey participants who did not experience an early dropout at the third question and who completed the survey without any dropouts. To achieve this, we utilize the drop_detect function, which identifies participants according to dropout status at specified points within the survey. By merging the output with our original data, we can then apply a filter to retain only the desired respondents. Once filtered, we remove the additional columns introduced by `drop_detect` as they are no longer necessary for further analysis. While not demonstrated in this example, this method of indexing is particularly advantageous when we need to perform complex manipulations, such as excluding all participants who dropped out before reaching the tenth item in the survey.
 
-    flying_dropouts <- flying %>% 
-    drop_detect(last_col = "location_census_region")
 
-    head(flying_dropouts)
+```r
+flying_dropouts <- flying %>% 
+drop_detect(last_col = "location_census_region")
 
-    flying_cleaned <- flying_dropouts %>% 
-    bind_cols(flying) %>% 
-    filter(dropout_col != "seat_recline" | dropout == FALSE) %>% 
-    select(-starts_with("dropout")) 
+head(flying_dropouts)
+
+flying_cleaned <- flying_dropouts %>% 
+bind_cols(flying) %>% 
[…]
