### 2023-12-20 #6070 @samhforbes
Thanks @EugeneHao for the prompt reply.

Sorry @max-alletsee I can see you have a new handle. Would you be willing to review this submission for JOSS?
