### 2024-01-02 #6070 @Darthpathos
Yes I’m also happy to review and the end of the month would be great - just
finishing a review for something else and can then get into this.  Please
send the pdf and I’ll get started asap.

View my Gift of 8 Campaign: http://beadonor.ca/christopher-battiston
