### 2024-01-02 #6070 @samhforbes
Hi @Darthpathos I would be happy to have you review as well - although please note it's the software as well as the PDF - see  https://joss.readthedocs.io/en/latest/review_criteria.html
If that's still OK with you I'll add your name, otherwise I will keep you in mind for future submissions!
