### 2024-01-02 #6070 @BlakeLW
Hi @samhforbes,
I am available to review the manuscript and the software in R. I can complete my review prior to January 26.
Please let me know how I should follow-up!
Best,
Blake
