### 2024-01-02 #6070 @Darthpathos
As I’m currently working on survey analysis in R, I would’ve reviewed the
software anyway :). looking forward to playing with the Package!

View my Gift of 8 Campaign: http://beadonor.ca/christopher-battiston
