### 2024-01-05 #6070 @Darthpathos
I’ve not heard back and wanted to follow up - I’m happy to do the review
and can have it done by the end of the month.

Thanks and have a great weekend!

View my Gift of 8 Campaign: http://beadonor.ca/christopher-battiston

On Tue, Jan 2, 2024 at 15:30 Christopher Battiston ***@***.***>
wrote:

> As I’m currently working on survey analysis in R, I would’ve reviewed the
> software anyway :). looking forward to playing with the Package!
>
> View my Gift of 8 Campaign: http://beadonor.ca/christopher-battiston
>
>
> On Tue, Jan 2, 2024 at 11:24 Sam Forbes ***@***.***> wrote:
>
>> Hi @Darthpathos <https://github.com/Darthpathos> I would be happy to
>> have you review as well - although please note it's the software as well as
>> the PDF - see https://joss.readthedocs.io/en/latest/review_criteria.html
>> If that's still OK with you I'll add your name, otherwise I will keep you
>> in mind for future submissions!
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/6070#issuecomment-1874246682>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ABYRG6IKYONUGN4TFCSEZT3YMQYEPAVCNFSM6AAAAAA7TIULVCVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTQNZUGI2DMNRYGI>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>
