### 2024-01-06 #6181 @editorialbot
[REVIEW]: dropout: an R Package for Addressing Dropouts, Missing Values, and Sectional Challenges in Survey Data Analysis
**Submitting author:** @hendr1km (<a href="http://orcid.org/0009-0007-3222-899X">Hendrik Mann</a>)
**Repository:** https://github.com/hendr1km/dropout
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.1
**Archive:** 10.6084/m9.figshare.25355746.v1
