### 2024-01-08 #6181 @Darthpathos
View my Gift of 8 Campaign: http://beadonor.ca/christopher-battiston

On Sat, Jan 6, 2024 at 08:43 The Open Journals editorial robot <
***@***.***> wrote:

> *Submitting author:* @hendr1km <https://github.com/hendr1km> (Hendrik Mann
> <http://orcid.org/0009-0007-3222-899X>)
> *Repository:* https://github.com/hendr1km/dropout
> *Branch with paper.md* (empty if default branch):
> *Version:* v2.1.0
> *Editor:* @samhforbes <https://github.com/samhforbes>
> *Reviewers:* @medewitt <https://github.com/medewitt>, @max-alletsee
> <https://github.com/max-alletsee>, @Darthpathos
> <https://github.com/Darthpathos>
> *Archive:* Pending
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/721cbec4ff9ab3a56a6ad9782d296ac6>
>
> Status badge code:
>
>
> *Reviewers and authors*:
>
> Please avoid lengthy details of difficulties in the review thread.
> Instead, please create a new issue in the target repository
> <https://github.com/hendr1km/dropout> and link to those issues
> (especially acceptance-blockers) by leaving comments in the review thread
> below. (For completists: if the target issue tracker is also on GitHub,
> linking the review thread in the issue or vice versa will create
> corresponding breadcrumb trails in the link target.)
> Reviewer instructions & questions
>
> @medewitt <https://github.com/medewitt> & @max-alletsee
> <https://github.com/max-alletsee> & @Darthpathos
> <https://github.com/Darthpathos>, your review will be checklist based.
> Each of you will have a separate checklist that you should update when
> carrying out your review.
> First of all you need to run this command in a separate comment to create
> the checklist:
>
> @editorialbot generate my checklist
>
> The reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any
> questions/concerns please let @samhforbes <https://github.com/samhforbes>
> know.
>
> ✨ *Please start on your review when you are able, and be sure to complete
> your review in the next six weeks, at the very latest* ✨
> Checklists
>
> @medewitt <https://github.com/medewitt>, please create your checklist
> typing: @editorialbot generate my checklist
>
> @max-alletsee <https://github.com/max-alletsee>, please create your
> checklist typing: @editorialbot generate my checklist
>
> @Darthpathos <https://github.com/Darthpathos>, please create your
> checklist typing: @editorialbot generate my checklist
>
