### 2024-01-09 #6181 @danielskatz
@Darthpathos - Your checklist is in the issue: see https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1882002143
