### 2024-02-21 hendr1km/dropout#2 @medewitt
Unable to install on R <= 4.0
## Problem

Unable to install on R version 3.6

## Solution

Two potential solutions

1. Update the minimum R version support to >= 4.0.0
2. Remove the use of the native pipe (`|>`) (e.g., https://github.com/hendr1km/dropout/blob/f125589d5f00435cc7c3c6feffc055b895ed51fe/R/drop_detect.R#L30)

As part of [the joss review](https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1958254503)
