### 2024-02-21 hendr1km/dropout#3 @medewitt
Code of Conduct per JOSS
## Issue

Lacking code of conduct per JOSS software standards

Part of [this review](https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1958254503)

>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
