### 2024-02-21 hendr1km/dropout#5 @medewitt
Joss Review: Position package in context with commonly used packages
## Issue

One of the JOSS criteria is describing the other relevant packages and what your package solves.

I would position your package within the commonly mentioned and used packaged for missing data in surveys [shown in the Taskview](https://cran.r-project.org/web/views/MissingData.html)

For example the [mde package](https://cran.r-project.org/web/packages/mde/vignettes/mde_vignette.html) gives nearly identical statistics to your package with some additional functionality (e.g., recoding and replacement)

>Missing data patterns can be identified and explored using the packages [mi](https://cran.r-project.org/web/packages/mi/index.html) (and its GUI [migui](https://cran.r-project.org/web/packages/migui/index.html)), [wrangle](https://cran.r-project.org/web/packages/wrangle/index.html), [DescTools](https://cran.r-project.org/web/packages/DescTools/index.html), and [naniar](https://cran.r-project.org/web/packages/naniar/index.html). [daqapo](https://cran.r-project.org/web/packages/daqapo/index.html) is a generic data quality toolbox that can also be used to identify missing data. More specifically, [ggmice](https://cran.r-project.org/web/packages/ggmice/index.html) produces plots for the [mice](https://cran.r-project.org/web/packages/mice/index.html) imputation workflow and can be used for missing data exploration and evaluation of imputation quality.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1958254503)
