### 2024-02-21 hendr1km/dropout#6 @medewitt
Joss Review--Supported claim of scalability
>The dropout package’s scalability is enhanced by the integration of C++ code through the Rcpp package, developed by Eddelbuettel et al. (2023). This incorporation makes dropout particularly effective for analyzing large survey datasets.

Unless this is supported with experimental results, I wouldn't include the second sentence in the paper. I totally understand that the C++ implementation should provide some speed, but unless experimentally supported I wouldn't include it.

Part of [this joss review](https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1958254503)
