### 2024-02-22 hendr1km/dropout#7 @medewitt
JOSS review: Cpp code not documented
You will likely need to document your C++ code using the Rcpp conventions (see [this Rcpp guide](https://gallery.rcpp.org/articles/documenting-rcpp-packages/) using Roxygen).

>Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Part of [this joss review](https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1958254503)
