### 2024-02-22 3d69707
removing the native pipe to allow lower R dep. Adding dplyr {{}} eval to remove depr warndings. Adding readme entry.
