+++ articles/paper.bib
@@ -1,3 +1,31 @@
+  @Manual{daqapo,
+    title = {daqapo: Data Quality Assessment for Process-Oriented Data},
+    author = {Niels Martin},
+    year = {2022},
+    note = {R package version 0.3.2},
+    url = {https://CRAN.R-project.org/package=daqapo},
+  }
+  
+   @Article{naniar,
+    title = {Expanding Tidy Data Principles to Facilitate Missing Data Exploration, Visualization and Assessment of Imputations},
+    author = {Nicholas Tierney and Dianne Cook},
+    journal = {Journal of Statistical Software},
+    year = {2023},
+    volume = {105},
+    number = {7},
+    pages = {1--31},
+    doi = {10.18637/jss.v105.i07},
+  }
+
+ 
+ @Manual{mde,
+    title = {mde: Missing Data Explorer},
+    author = {Nelson Gonzabato},
+    year = {2022},
+    note = {R package version 0.3.2},
+    url = {https://CRAN.R-project.org/package=mde},
+  }
+
 @article{Pearson:2017,
   	url = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
   	Archiveprefix = {arXiv},
