+++ articles/paper.md
@@ -29,6 +29,17 @@ Unlike current R packages that deal with missing values in data analysis by (vis
 However, unlike `daqapo` [@daqapo], a package that can also be used to find sequential gaps of missing values for participants, `dropout` allows you to use its summary statistics and detect participants based on their classified response pattern in your further data cleaning and analysis workflow. 
 
 
+# Installation
+The dropout package can be installed using CRAN, as well as GitHub for access to the development version. For the second type of installation devtools is needed.
+```r
+# CRAN
+install.packages("dropout")
+
+# GitHub 
+install.packages("devtools")
+devtools::install_github("hendr1km/dropout")
+```
+
 # Usage
 
 `dropout` includes two essential functions, beginning with `drop_summary`. This function is designed to generate summary statistics for missing values in your dataset. It requires two primary inputs: the dataframe to analyze and the identifier of the last survey item. The latter is particularly important in cases where your data includes additional columns generated by the survey platform, such as participation time. These extra items can lead to biased results in the function, as no dropouts may be detected if the last column has no missing values. To counter this, if the last_col argument is left unspecified, `drop_summary` will automatically address the issue by setting `last_col` to the last column in the dataframe that contains missing values, and it will issue a warning to inform the user. For optimal usage of the `dropout` package, it's generally good practice to ensure your dataframe either exclusively contains the survey items, or that the `last_col` argument is manually set to the last survey item in your dataset. Generally, it is essential to index all survey items in the dataframe in their correct order. An optional argument of the `drop_summary` function is `section_min`. This parameter plays a crucial role when `drop_summary` is used to detect sections that participants may have skipped. By default, it looks for at least three consecutive missing items to identify a skipped section. This threshold seems to be sensitive enough to differentiate between single missing values and section omissions. However, it's advisable to experiment with different `section_min` settings to find the optimal threshold that aligns with your specific study design. 
