### 2024-02-22 hendr1km/dropout#2 @hendr1km
First of all thank you for all your feedback.

Regarding this issue I have removed the native pipe from the code. In the test run the updated dev version worked now with R 3.6
