### 2024-02-22 hendr1km/dropout#7 @hendr1km
I have added a documentation of the c++ Code following the guide
