### 2024-02-22 hendr1km/dropout#6 @hendr1km
I have removed this part from the paper
