### 2024-02-22 84572aa
removing start up message
