+++ R/zzz.R
@@ -1,4 +1,0 @@
-.onAttach <- function(libname, pkgname) {
- packageStartupMessage("This is version ", packageVersion(pkgname),
-                       " of ", pkgname,". \n Please report any issues at https://github.com/hendr1km/dropout/issues")
-}
