### 2024-02-22 hendr1km/dropout#3 @hendr1km
I have added guidelines to the repo
