### 2024-02-24 hendr1km/dropout#8 @max-alletsee
JOSS Review: Citation of flying data
One of the JOSS criteria is a complete list of references, including datasets.

While you are using a dataset that comes with the package installation and you provide a citation in the package documentation, I would encourage you to also cite it directly in the software paper.

(As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1880197553).)
