### 2024-02-24 hendr1km/dropout#11 @max-alletsee
JOSS Review: minor typo in paper
There is a minor typo in the last paragraph of the software paper (line 103 in the PDF proof):

> In the **follwing** example of Figure 2, we utilize the drop_summary function...

... should read as "**following**"

(As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6181#issuecomment-1880197553).)
