### 2024-02-28 hendr1km/dropout#2 @medewitt
Excellent! Last thing is just to change your DESCRIPTION to reflect R >= 3.5

https://github.com/hendr1km/dropout/blob/84572aa8d33d6d35430d4a43fec6db71cd8217b3/DESCRIPTION#L25

That's the minimum R version supported by dplyr.
