### 2024-02-29 hendr1km/dropout#8 @hendr1km
Thank you for your feedback. I have added the citation to the paper.
