+++ DESCRIPTION
@@ -22,7 +22,7 @@ LinkingTo:
     Rcpp
 Config/testthat/edition: 3
 Depends: 
-    R (>= 2.10),
+    R (>= 3.5),
 LazyData: true
 VignetteBuilder: knitr
 
