### 2024-02-29 21b421d
fixing a bug that loads to the R session crashing when section_min is se to 0
