+++ articles/paper.md
@@ -28,18 +28,6 @@ Missing data can reduce the statistical power of a study and can produce biased 
 Unlike current R packages that deal with missing values in data analysis by (visually) producing summary statistics that treat each missing value equally, such as `naniar` [@naniar] and `mde` [@mde], `dropout` classifies missing values based on the occurrence of single missing values, section dropouts and complete dropouts. This allows `dropout` to produce summary statistics of different response patterns and relate them to the overall occurrence of missing values.
 Unlike `daqapo` [@daqapo], a package that can also be used to find sequential gaps of missing values for participants, `dropout` allows you to use its summary statistics and detect participants based on their classified response pattern. This enables a simpler workflow for data cleaning and further analysis.
 
-
-# Installation
-The dropout package can be installed using CRAN, as well as GitHub for access to the development version. For the second type of installation devtools is needed.
-```r
-# CRAN
-install.packages("dropout")
-
-# GitHub 
-install.packages("devtools")
-devtools::install_github("hendr1km/dropout")
-```
-
 # Usage
 
 `dropout` includes two essential functions, beginning with `drop_summary`. This function is designed to generate summary statistics for missing values in your dataset. It requires two primary inputs: the dataframe to analyze and the identifier of the last survey item. The latter is particularly important in cases where your data includes additional columns generated by the survey platform, such as participation time. These extra items can lead to biased results in the function, as no dropouts may be detected if the last column has no missing values. To counter this, if the last_col argument is left unspecified, `drop_summary` will automatically address the issue by setting `last_col` to the last column in the dataframe that contains missing values, and it will issue a warning to inform the user. For optimal usage of the `dropout` package, it's generally good practice to ensure your dataframe either exclusively contains the survey items, or that the `last_col` argument is manually set to the last survey item in your dataset. Generally, it is essential to index all survey items in the dataframe in their correct order. An optional argument of the `drop_summary` function is `section_min`. This parameter plays a crucial role when `drop_summary` is used to detect sections that participants may have skipped. By default, it looks for at least three consecutive missing items to identify a skipped section. This threshold seems to be sensitive enough to differentiate between single missing values and section omissions. However, it's advisable to experiment with different `section_min` settings to find the optimal threshold that aligns with your specific study design. 
@@ -48,7 +36,7 @@ devtools::install_github("hendr1km/dropout")
 
 # Examples
 
-For the following examples we will use an adapted version of the Flying Etiquette by five-thirty-eight dataset that is included in the `dropout` package. In these workflow examples, I will be using dplyr verbs [@dplyr] - although this is not necessary.
+For the following examples we will use an adapted version of the Flying Etiquette by five-thirty-eight dataset [@flying] that is included in the `dropout` package. In these workflow examples, I will be using dplyr verbs [@dplyr] - although this is not necessary.
 
 As illustrated with the `flying` dataset example, even though all columns are arranged in the correct order of survey items, the last column survey_type does not correspond to a survey item. In such scenarios, the dropout package intuitively addresses this issue by disregarding the non-survey column and automatically setting the last_col argument to `location_census_region`. This adjustment is accompanied by a warning to inform the user. However, in more complex situations, it's advisable to either create a subset of your data or manually set the last_col argument to the actual last survey item. We will demonstrate this approach in the following examples.
 
@@ -132,7 +120,7 @@ pivot_longer(3:4, values_to = "n", names_to = "condition") %>%
 ```
 
 
[…]
