+++ articles/paper.bib
@@ -1,3 +1,16 @@
+@online{flying,
+  author  = {Walt Hickey},
+  title   = {41 Percent Of Fliers Think You’re Rude If You Recline Your Seat},
+  year    = {2014},
+  month   = sep,
+  day     = {6},
+  publisher = {GitHub},
+  url     = {https://github.com/fivethirtyeight/data/tree/master/flying-etiquette-survey},
+  urldate = {2024-02-29}, % Date you last accessed the URL
+  note    = {Flying Etiquette Survey [Data set]}
+}
+
+
   @Manual{daqapo,
     title = {daqapo: Data Quality Assessment for Process-Oriented Data},
     author = {Niels Martin},
@@ -5,7 +18,7 @@
     note = {R package version 0.3.2},
     url = {https://CRAN.R-project.org/package=daqapo},
   }
-  
+
    @Article{naniar,
     title = {Expanding Tidy Data Principles to Facilitate Missing Data Exploration, Visualization and Assessment of Imputations},
     author = {Nicholas Tierney and Dianne Cook},
@@ -17,7 +30,7 @@
     doi = {10.18637/jss.v105.i07},
   }
 
- 
+
  @Manual{mde,
     title = {mde: Missing Data Explorer},
     author = {Nelson Gonzabato},
@@ -116,7 +129,7 @@
     note = {R package version 1.0.11},
     url = {https://CRAN.R-project.org/package=Rcpp},
   }
-  
+
 @Article{tidyverse,
     title = {Welcome to the {tidyverse}},
     author = {Hadley Wickham and Mara Averick and Jennifer Bryan and
