### 2024-03-04 #6181 @max-alletsee
@samhforbes I saw that @hendr1km made all changes that I suggested. I just added the last missing checkbox for citations.
