### 2024-03-06 a33e144
change descr version for cran
