+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: dropout
 Title: Handling Incomplete Responses in Survey Data Analysis
-Version: 2.1.0
+Version: 2.1.1
 Authors@R: c(person(given = "Hendrik", family = "Mann", email = "hendrik.mann@uni-wuppertal.de", role = c("aut", "cre")))
 Description: Offers robust tools to identify and manage incomplete responses in survey datasets, thereby enhancing the quality and reliability of research findings.
 License: MIT + file LICENSE
