+++ articles/paper.bib
@@ -2,11 +2,9 @@
   author  = {Walt Hickey},
   title   = {41 Percent Of Fliers Think You’re Rude If You Recline Your Seat},
   year    = {2014},
-  month   = sep,
   day     = {6},
   publisher = {GitHub},
   url     = {https://github.com/fivethirtyeight/data/tree/master/flying-etiquette-survey},
-  urldate = {2024-02-29}, % Date you last accessed the URL
   note    = {Flying Etiquette Survey [Data set]}
 }
 
