### 2024-03-18 #6181 @samhforbes
@hendr1km  The only minor things for me are:
- line 95 in method 2 rather than the method 2
- I'm fairly certain the DOI readout is correct (no missing DOIs that I can see where it's flagged them), but please double check.
