### 2024-03-18 34bd7c4
Update paper.md
minor article adjustment
