+++ articles/paper.md
@@ -74,7 +74,7 @@ select(-starts_with("dropout"))
 ```
 
 
-Next, we aim to exclude the 164 participants who skipped an entire section of the survey without fully dropping out. This can be accomplished through two approaches. The first method involves setting the `last_col` argument of the `drop_detect` function to the last column of the omitted section. By doing so, all participants who skipped the entire section will be flagged as dropouts, making it straightforward to exclude them. The second method requires creating a subset of the dataset that includes only the columns of the concerned section. This subset can then be used to specifically filter for section-based dropouts. It is important to note that when working with such a subset, the indices provided by `dropout_index` might correspond to the subset's dataframe and not the original one. This distinction is crucial for accurately mapping the dropout information back to the complete dataset, when using a subset starting not from column one of the origional dataset. In the method 2 this is not an issue.  
+Next, we aim to exclude the 164 participants who skipped an entire section of the survey without fully dropping out. This can be accomplished through two approaches. The first method involves setting the `last_col` argument of the `drop_detect` function to the last column of the omitted section. By doing so, all participants who skipped the entire section will be flagged as dropouts, making it straightforward to exclude them. The second method requires creating a subset of the dataset that includes only the columns of the concerned section. This subset can then be used to specifically filter for section-based dropouts. It is important to note that when working with such a subset, the indices provided by `dropout_index` might correspond to the subset's dataframe and not the original one. This distinction is crucial for accurately mapping the dropout information back to the complete dataset, when using a subset starting not from column one of the origional dataset. In method 2 this is not an issue.  
 
 ```r
 # method 1 (recommended as indexes of dropout_index will still match the data)
