### 2024-03-18 #6181 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
