### 2024-03-25 6ac9868
Update paper.md
Minor, but not a hyphen: it's an m-dash.
