### 2024-03-30 #6181 @oliviaguest
Huge thanks to the reviewers @medewitt, @max-alletsee, @Darthpathos and editor @samhforbes! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors, @hendr1km! 🥳 🍾
